refactor: co-locate setup checks into system/setup.go
Moves Incompatibility and CheckIncompatibilities out of the unix-only
docker_unix.go alongside the missing-dependency helpers, renaming
dependencies.go to setup.go. Adds a Windows stub for IsDockerUsingLibkrun
so CheckIncompatibilities is now platform-neutral, fixing a latent
Windows build failure in cmd/project.
