+++ internal/system/setup_test.go
@@ -1,0 +1,15 @@
+package system
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func TestCheckIncompatibilities(t *testing.T) {
+	t.Run("no incompatibilities on non-darwin", func(t *testing.T) {
+		t.Setenv("HOME", "/tmp/test-home")
+		incompatibilities := CheckIncompatibilities(false, "/tmp/project")
+		assert.Empty(t, incompatibilities)
+	})
+}
